requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.21
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
requests==2.31.0
beautifulsoup4==4.12.2
twilio==8.9.1
gunicorn==21.2.0
qrcode==7.4.2
Pillow==10.0.1
schedule==1.2.0
python-dotenv==1.0.0
flask-cors==4.0.0
plotly==5.16.1
openpyxl==3.1.2
python-dateutil==2.8.2
pytz==2023.3
charset-normalizer==3.3.0
urllib3==2.0.5
certifi==2023.7.22
jinja2==3.1.2
werkzeug==2.3.7
markupsafe==2.1.3
click==8.1.7
blinker==1.6.2
itsdangerous==2.1.2
typing-extensions==4.8.0 